Add files via upload
Added the final (bonus) simulation experiment which compares the methods in the case where the target species' abundance decreases over time. Separately considers the case where the main competitor's abundance increases and decreases. Uploaded are both the .R file used to generate the results and the .rds file with the results.